change to double

+++ features/merchant/shop_page/src/main/java/com/tokopedia/shop/analytic/ShopPageHomeTracking.kt
@@ -2760,7 +2760,7 @@ class ShopPageHomeTracking(
             putString(ITEM_ID, shopHomeProductUiModel.id)
             putString(ITEM_NAME, shopHomeProductUiModel.name)
             putString(ITEM_VARIANT, "")
-            putLong(PRICE, shopHomeProductUiModel.displayedPrice.digitsOnly())
+            putDouble(PRICE, shopHomeProductUiModel.displayedPrice.digitsOnly().toDouble())
         }
     }
